% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_cost.R
\name{monthly_cost}
\alias{monthly_cost}
\title{Calculate monthly cost of a house}
\usage{
monthly_cost(
  house_value = 20 * 10000,
  property_tax_rate = 1.54/100,
  other_tax = 1/100,
  one_time_fee = 1000,
  annual_other_tax = 3000,
  monthly_hoa_fee = 150,
  monthly_home_insurance = 150,
  monthly_utility_fee = 300,
  closing_fee,
  year = 5,
  saving_rate = 0.03,
  down_payment_rate = 0.2,
  mortgage_rate = 0.0375,
  morrgage_year = 30
)
}
\arguments{
\item{house_value}{house value}

\item{property_tax_rate}{annual property tax rate}

\item{other_tax}{annual other tax rate}

\item{one_time_fee}{one time fee, e.g., agent fee, points fee}

\item{annual_other_tax}{other tex fee, e.g., school/county tax}

\item{monthly_hoa_fee}{monthly HOA fee}

\item{monthly_home_insurance}{monthly home insurance fee}

\item{monthly_utility_fee}{monthly utility fee}

\item{closing_fee}{closing fee}

\item{year}{number of years for invest this house}

\item{saving_rate}{saving rate, which is potential loss of the down payment}

\item{down_payment_rate}{down payment}

\item{mortgage_rate}{annual mortgage rate}

\item{morrgage_year}{mortgage years}
}
\description{
Calculate monthly cost of a house
}
