% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repayment.R
\name{repayment}
\alias{repayment}
\title{Calculate the monthly mortgage}
\usage{
repayment(A, beta, m)
}
\arguments{
\item{A}{loan principal}

\item{beta}{monthly loan rate}

\item{m}{number of periods in months}
}
\description{
Calculate the monthly mortgage
}
